requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-apache-spark==4.1.3
airflow-provider-great-expectations==0.2.6
great-expectations==0.17.11
minio==7.1.16
clickhouse-connect==0.6.10